Move ts abstraction to tsapi.ts
