+++ lib/project.ts
@@ -1,6 +1,7 @@
 ///<reference path='../definitions/ref.d.ts'/>
 
 import ts = require('typescript');
+import tsApi = require('./tsapi');
 import main = require('./main');
 import gutil = require('gulp-util');
 import sourceMap = require('source-map');
@@ -27,30 +28,7 @@ interface OutputFile {
 	sourcemap?: Object;
 }
 
-/*
- * TS1.4 had a simple getDiagnostics method, in 1.5 that method doens't exist,
- * but instead there are now 4 methods which (combined) return all diagnostics.
- */
-interface Program14 {
-	getDiagnostics(): ts.Diagnostic[];
-	getTypeChecker(fullTypeCheckMode: boolean): ts.TypeChecker;
-}
-interface Program15 {
-	getSyntacticDiagnostics(): ts.Diagnostic[];
-	getGlobalDiagnostics(): ts.Diagnostic[];
-	getSemanticDiagnostics(): ts.Diagnostic[];
-	getDeclarationDiagnostics(): ts.Diagnostic[];
-	emit(): { diagnostics: ts.Diagnostic[]; };
-}
-/*
- * In TS 14 the method getLineAndCharacterFromPosition has been renamed from ...From... to ...Of...
- */
-interface TSFile14 {
-	getLineAndCharacterFromPosition(pos: number): ts.LineAndCharacter;
-}
-interface TSFile15 {
-	getLineAndCharacterOfPosition(pos: number): ts.LineAndCharacter;
-}
+
 
 export class Project {
 	static unresolvedFile: FileData = {
@@ -59,47 +37,6 @@ export class Project {
 		content: undefined,
 		ts: undefined
 	};
-
-	static getFileName(thing: { filename: string} | { fileName: string }): string {
-		if ((<any> thing).filename) return (<any> thing).filename;
-		return (<any> thing).fileName;
-	}
-	private static getDiagnosticsAndEmit(program: Program14 | Program15): ts.Diagnostic[] {
-		if ((<Program14> program).getDiagnostics) {
-			var errors = (<Program14> program).getDiagnostics();
-
-			if (!errors.length) {
-				// If there are no syntax errors, check types
-				var checker = (<Program14> program).getTypeChecker(true);
-
-				var semanticErrors = checker.getDiagnostics();
-
-				var emitErrors = checker.emitFiles().diagnostics;
-
-				errors = semanticErrors.concat(emitErrors);
-			}
-
-			return errors;
-		} else {
-			var errors = (<Program15> program).getSyntacticDiagnostics();
-			if (errors.length === 0) errors = (<Program15> program).getGlobalDiagnostics();
-			if (errors.length === 0) errors = (<Program15> program).getSemanticDiagnostics();
-
-			var emitOutput = (<Program15> program).emit();
-			return errors.concat(emitOutput.diagnostics);
-		}
-	}
-	private static getLineAndCharacterOfPosition(typescript: typeof ts, file: TSFile14 | TSFile15, position: number) {
-		if ((<TSFile15> file).getLineAndCharacterOfPosition) { // TS 1.5
-			var lineAndCharacter = (<TSFile15> file).getLineAndCharacterOfPosition(position);
-			return {
-				line: lineAndCharacter.line + 1,
-				character: lineAndCharacter.character + 1
-			}
-		} else { // TS 1.4
-			return (<TSFile14> file).getLineAndCharacterFromPosition(position);
-		}
-	}
 
 	/**
 	 * The TypeScript library that is used for this project.
@@ -233,7 +170,7 @@ export class Project {
 			return err;
 		}
 
-		var filename = this.getOriginalName(Project.getFileName(info.file));
+		var filename = this.getOriginalName(tsApi.getFileName(info.file));
 		var file = this.host.getFileData(filename);
 
 		if (file) {
@@ -247,12 +184,12 @@ export class Project {
 				filename = file.originalFilename;
 			}
 		} else {
-			filename = Project.getFileName(info.file);
+			filename = tsApi.getFileName(info.file);
 			err.fullFilename = filename;
 		}
 
-		var startPos = Project.getLineAndCharacterOfPosition(this.typescript, info.file, info.start);
-		var endPos = Project.getLineAndCharacterOfPosition(this.typescript, info.file, info.start + info.length);
+		var startPos = tsApi.getLineAndCharacterOfPosition(this.typescript, info.file, info.start);
+		var endPos = tsApi.getLineAndCharacterOfPosition(this.typescript, info.file, info.start + info.length);
 
 		err.startPosition = {
 			position: info.start,
@@ -322,7 +259,7 @@ export class Project {
 	}
 
 	private resolve(session: { tasks: number; callback: () => void; }, file: FileData) {
-		var references = file.ts.referencedFiles.map(item => path.join(path.dirname(Project.getFileName(file.ts)), Project.getFileName(item)));
+		var references = file.ts.referencedFiles.map(item => path.join(path.dirname(tsApi.getFileName(file.ts)), tsApi.getFileName(item)));
 
 		this.typescript.forEachChild(file.ts, (node) => {
 			if (node.kind === (<any> this.typescript.SyntaxKind).ImportDeclaration) {
@@ -338,7 +275,7 @@ export class Project {
 				if (typeof (<ts.StringLiteralExpression> reference).text !== 'string') {
 					return;
 				}
-				var ref = path.join(path.dirname(Project.getFileName(file.ts)), (<ts.StringLiteralExpression> reference).text);
+				var ref = path.join(path.dirname(tsApi.getFileName(file.ts)), (<ts.StringLiteralExpression> reference).text);
 
 				// Don't know if this name is defined with `declare module 'foo'`, but let's load it to be sure.
 				// We guess what file the user wants. This will be right in most cases.
@@ -430,7 +367,7 @@ export class Project {
 		// Creating a program compiles the sources
 		this.program = this.typescript.createProgram(rootFilenames, this.options, this.host);
 
-		var errors = Project.getDiagnosticsAndEmit(this.program);
+		var errors = tsApi.getDiagnosticsAndEmit(this.program);
 
 		for (var i = 0; i < errors.length; i++) {
 			errorCallback(this.getError(errors[i]));
@@ -552,7 +489,7 @@ export class Project {
 
 				var inputFile = this.currentFiles[originalName];
 				var tsFile = this.program.getSourceFile(originalName);
-				var references = tsFile.referencedFiles.map(file => Project.getFileName(file));
+				var references = tsFile.referencedFiles.map(file => tsApi.getFileName(file));
 
 				for (var j = 0; j < outputJS.length; ++j) {
 					var other = outputJS[j];
